requirements.txt
anyio==3.7.0
astroid==2.15.5
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dill==0.3.6
exceptiongroup==1.1.1
fastapi==0.96.0
h11==0.14.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==3.5.1
pydantic==1.10.9
pylint==2.17.4
PyYAML==6.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
wrapt==1.15.0
